pyproject.toml
[project]
name = "sistema-de-pedido-lanchonete"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.122.0",
    "flake8>=7.3.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pylint>=4.0.4",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "tzdata",
]

[tool.ruff]
line-length = 80
fix = true


[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev .\main.py'
test =  'pytest -s -x --cov=. -vv'
post_test = 'coverage html'
fix = 'ruff check . --fix'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'


[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.5",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
    "taskipy>=1.14.1",
]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring (Não exigir docs no topo de todo arquivo)
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "W0621",  # redefined-outer-name (Essencial desativar se usa Pytest/fixtures)
    "R0903",  # too-few-public-methods (FastAPI usa muitos modelos simples)
    "R0801"   # similar-lines (muito pesado)
]

[tool.pylint.main]
jobs = 4 # Usa múltiplos núcleos (muito mais rápido)
persistent = false 


[tool.pylint.format]
max-line-length = 140
